Wait for file writing threads before we take the files
Intermittent failures were seen on Windows, where the file was in the
middle of being written when we copied it. Waiting for the threads
ensures that they are done before we take the files.

We need to wait on the thread pool to drain rather than using
`stop_and_drain` because otherwise the destructor of the writer
will post to the stopped thread pool. This will throw an exception
and the `noxcept nature of the destructor would result in a call to
`std::terminate`.