Bugfix using `pod5 recover` on multiple files
Without this fix the recovery process tries to recover all files to the
same destination, based on the last iteration of the preceding loop.
With this fix all files being recovered get written to separate output
files.

I have confirmed the behaviour before and after this commit using manual
testing.