Bugfix `POD5_PYTHON_RETURN_NOT_OK` macro
The argument passed to this macro would previously be evaluated twice in
the case where an error message is returned. This is particularly
unhelpful in the case where side effects change the error message the
second time around.

Using a do ... while(false) loop in the macro forces invocations of
this statement like macro to be followed by statement terminating `;`.