Suppress FileNotFoundError instead of checking for existence
If the file has already been removed then being unable to remove it
isn't a problem. Therefore it is reasonable to ignore the error.
Ignoring the error is better than checking for the existence as it
removes the possibility of another process removing it between the call
to `exists` and `unlink`.