[INSTX-9693] Significantly improve performance of the file fuzzer
Unlike |std::vector| which amortizes its growth, |builder.Append()|
resizes by 1 element at a time. This means that the majority of the
fuzzer's time is spent in ASAN reallocating.

The builder's base class has an |AppendScalar()| that takes a repeat
count, however that isn't available for raw types. Passing a pair of
iterators is allowed though so we'll use that.

This takes us from ~30 execs/s to >800 execs/s locally.